Moves version module
