+++ lib/recurly.js
@@ -3,13 +3,13 @@
 import clone from 'component-clone';
 import Emitter from 'component-emitter';
 import uuid from 'uuid/v4';
-import errors from './errors';
-import version from './version';
+import errors from './recurly/errors';
 import bankAccount from './recurly/bank-account';
 import coupon from './recurly/coupon';
 import giftcard from './recurly/giftcard';
 import plan from './recurly/plan';
 import tax from './recurly/tax';
+import version from './recurly/version';
 import {request, cachedRequest, pipedRequest} from './recurly/request';
 import {token} from './recurly/token';
 import {factory as Adyen} from './recurly/adyen';
